requirements.txt
absl-py==1.4.0
aiobotocore==2.12.1
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.2.0
anyio==3.7.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
botocore==1.34.51
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
cryptography==42.0.5
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
importlib-metadata==6.11.0
importlib-resources==5.13.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
packaging==23.2
pandas==1.5.3
Pillow==9.5.0
platformdirs==3.11.0
protobuf==4.25.3
pyarrow==15.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.17.2
PyJWT==2.8.0
Pympler==1.0.1
pyOpenSSL==24.0.0
python-dateutil==2.9.0.post0
pytimeparse==1.1.8
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.0
s3fs==2023.12.2
scikit-learn==1.3.2
scipy==1.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowflake-connector-python==3.7.1
snowflake-ml-python==1.2.3
snowflake-snowpark-python==1.13.0
sortedcontainers==2.4.0
sqlparse==0.4.4
streamlit==1.26.0
tenacity==8.2.3
threadpoolctl==3.3.0
toml==0.10.2
tomlkit==0.12.4
toolz==0.12.1
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==4.3.1
urllib3==1.26.18
validators==0.22.0
wrapt==1.16.0
xgboost==1.7.6
yarl==1.9.4
zipp==3.17.0
